tests: rust: extract common code
